+++ ChangeLog.md
@@ -3,13 +3,16 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
-* 2017/02/?? - 1.61 - PENDING
+* 2017/02/09 - 1.62
     * Renamed 'Mapping Qualifier' field to 'Equivalence Type' (Jazz: 440248)
     * Now pulling Equivalence Type' mapping item field options from the appropriate concepts (Jazz: 440248)
     * Added option to UIHelper.createSelectFieldString to create empty option tags
     * Fixed autosuggest field so if you clear the field completely it will clear out the underlying value
     * Renamed some map set fields and pulled IDs from metadata instead of hardcoding.
     * Added 'Source' or 'Target' to mapset calculated column labels
+
+* 2017/02/08 - 1.61
+	* Intermediate test build
 
 * 2017/02/03 - 1.60
     * Updated mapping to allow item field order to be chosen and calculated fields to be added to map items. Both occur only on mapset create. (Jazz: 444802, 430117)